+++ src/Compute/Compute/Strategies/ComputeRp/VirtualMachineScaleSetStrategy.cs
@@ -54,7 +54,11 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             IEnumerable<int> dataDisks,
             IList<string> zones,
             bool ultraSSDEnabled,
-            Func<IEngine, SubResource> proximityPlacementGroup)
+            Func<IEngine, SubResource> proximityPlacementGroup,
+            string priority,
+            string evictionPolicy,
+            double? maxPrice,
+            string[] scaleInPolicy)
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
                 name: name,
@@ -115,9 +119,16 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                                     NetworkSecurityGroup = engine.GetReference(networkSecurityGroup)
                                 }
                             }
-                        }
+                        },
+                        Priority = priority,
+                        EvictionPolicy = evictionPolicy,
+                        BillingProfile = (maxPrice == null) ? null : new BillingProfile(maxPrice)
                     },
                     ProximityPlacementGroup = proximityPlacementGroup(engine),
+                    ScaleInPolicy = (scaleInPolicy == null) ? null : new ScaleInPolicy
+                    {
+                        Rules = scaleInPolicy
+                    }
                 });
     }
 }