+++ src/Compute/Compute/ChangeLog.md
@@ -19,6 +19,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Breaking changes
+    - UploadSizeInBytes parameter is used instead of DiskSizeGB for New-AzDiskConfig when CreateOption is Upload
+    - PublishingProfile.Source.ManagedImage.Id is replaced with StorageProfile.Source.Id in GalleryImageVersion object
 
 ## Version 2.7.0
 * Add Priority, EvictionPolicy, and MaxPrice parameters to New-AzVM and New-AzVmss cmdlets