+++ src/Compute/Compute/ChangeLog.md
@@ -19,6 +19,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Add PublicIPAddressVersion parameter to New-AzVmssIPConfig
 * Move FileUris of custom script extension from public setting to protected setting
 * Add ScaleInPolicy to New-AzVmss, New-AzVmssConfig and Update-AzVmss cmdlets
 * Breaking changes